[REFACTOR] CommunityControllerTests -> JcoApplicationTests save관련 테스트 이동
